Remove "Acl" class from public API.
